Created a class to handle the ScrollToTop functionality for all the classes that need it. Finished implementing scroll functionality to PIVBasics1.
