+++ app/src/main/res/layout/activity_pivbasics1.xml
@@ -16,9 +16,9 @@
         android:layout_marginStart="175dp"
         android:background="#6D353232"
         android:backgroundTint="#6D353232"
-        android:visibility="gone"
         android:src="@drawable/ic_baseline_arrow_upward_24"
-        app:rippleColor="#6D353232" />
+        android:visibility="gone"
+        app:backgroundTint="#6D353232" />
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/nestedScroll"
